aidm_v3/requirements.txt
# LLM SDKs
anthropic>=0.40.0
google-generativeai
openai>=1.50.0
cryptography>=42.0.0

# Database
sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # Only needed for PostgreSQL
alembic>=1.13.0

# Web API
fastapi>=0.100.0
uvicorn>=0.24.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
ruff>=0.15.0
